Unify Audio- and instrument clip horizontal shift
