+++ src/deluge/model/action/action.h
@@ -61,7 +61,7 @@ class ModelStack;
 #define ACTION_ARRANGEMENT_TIME_CONTRACT 17
 #define ACTION_ARRANGEMENT_CLEAR 18
 #define ACTION_ARRANGEMENT_RECORD 19
-#define ACTION_INSTRUMENT_CLIP_HORIZONTAL_SHIFT 20
+#define ACTION_CLIP_HORIZONTAL_SHIFT 20
 #define ACTION_NOTE_NUDGE 21
 #define ACTION_NOTE_REPEAT_EDIT 22
 #define ACTION_EUCLIDEAN_NUM_EVENTS_EDIT 23
