+++ src/deluge/model/clip/instrument_clip.h
@@ -92,7 +92,8 @@ public:
 	void restoreBackedUpParamManagerMIDI(ModelStackWithModControllable* modelStack);
 	int getNoteRowId(NoteRow* noteRow, int noteRowIndex);
 	NoteRow* getNoteRowFromId(int id);
-	void shiftHorizontally(ModelStackWithTimelineCounter* modelStack, int amount);
+	/// Return true if successfully shifted. Instrument clips always succeed
+	bool shiftHorizontally(ModelStackWithTimelineCounter* modelStack, int amount);
 	bool containsAnyNotes();
 	ModelStackWithNoteRow* getNoteRowOnScreen(int yDisplay, ModelStackWithTimelineCounter* modelStack);
 	NoteRow* getNoteRowOnScreen(int yDisplay, Song* song, int* getIndex = NULL);
